refactor: 리더보드 Leader 타입 분리 및 import 구조 정리
