+++ types/leaderboard.ts
@@ -1,0 +1,9 @@
+export type Leader = {
+    rank: number;
+    name: string;
+    handle: string;
+    imageUrl: string;
+    wpm: number;
+    accuracy: number;
+    combo: number;
+  };
